+++ src/funcs/directoryDirectoryMemberDownload.ts
@@ -5,6 +5,7 @@
 import * as z from "zod";
 import { SyllableSDKCore } from "../core.js";
 import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
@@ -42,9 +43,9 @@ export async function directoryDirectoryMemberDownload(
 > {
   const path = pathToFunc("/api/v1/directory_members/download/")();
 
-  const headers = new Headers({
+  const headers = new Headers(compactMap({
     Accept: "application/json",
-  });
+  }));
 
   const secConfig = await extractSecurity(client._options.apiKeyHeader);
   const securityInput = secConfig == null ? {} : { apiKeyHeader: secConfig };