+++ src/funcs/toolsList.ts
@@ -5,6 +5,7 @@
 import { SyllableSDKCore } from "../core.js";
 import { encodeFormQuery } from "../lib/encodings.js";
 import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
 import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
@@ -35,7 +36,7 @@ export async function toolsList(
   options?: RequestOptions,
 ): Promise<
   Result<
-    components.ListResponseTool,
+    components.ListResponseToolResponse,
     | errors.HTTPValidationError
     | SDKError
     | SDKValidationError
@@ -71,9 +72,9 @@ export async function toolsList(
     "start_datetime": payload.start_datetime,
   });
 
-  const headers = new Headers({
+  const headers = new Headers(compactMap({
     Accept: "application/json",
-  });
+  }));
 
   const secConfig = await extractSecurity(client._options.apiKeyHeader);
   const securityInput = secConfig == null ? {} : { apiKeyHeader: secConfig };
@@ -123,7 +124,7 @@ export async function toolsList(
   };
 
   const [result] = await M.match<
-    components.ListResponseTool,
+    components.ListResponseToolResponse,
     | errors.HTTPValidationError
     | SDKError
     | SDKValidationError
@@ -133,7 +134,7 @@ export async function toolsList(
     | RequestTimeoutError
     | ConnectionError
   >(
-    M.json(200, components.ListResponseTool$inboundSchema),
+    M.json(200, components.ListResponseToolResponse$inboundSchema),
     M.jsonErr(422, errors.HTTPValidationError$inboundSchema),
     M.fail(["4XX", "5XX"]),
   )(response, { extraFields: responseFields });