+++ src/lib/config.ts
@@ -61,8 +61,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.2",
-  sdkVersion: "0.1.0-alpha.29",
-  genVersion: "2.492.3",
+  sdkVersion: "0.1.0-alpha.30",
+  genVersion: "2.493.13",
   userAgent:
-    "speakeasy-sdk/typescript 0.1.0-alpha.29 2.492.3 0.0.2 syllable-sdk",
+    "speakeasy-sdk/typescript 0.1.0-alpha.30 2.493.13 0.0.2 syllable-sdk",
 } as const;